pyproject.toml
[project]
name = "asx-stock-analysis"
version = "0.1.0"
description = "ASX stock price analysis and ML predictions using dbt and technical indicators"
requires-python = ">=3.11"
dependencies = [
    "dlt>=1.4.0",
    "dbt-duckdb>=1.10.0",
    "yfinance>=0.2.0",
    "duckdb>=1.0.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "apscheduler>=3.10.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "prophet>=1.1.0",
    "xgboost>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
asx-extract = "ingestion.asx_extraction.extract:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[dependency-groups]
dev = [
    "ruff>=0.4.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingestion"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function call in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["ingestion"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pre-commit]
repos = [
    { repo = "https://github.com/astral-sh/ruff-pre-commit", rev = "v0.4.0" },
    { repo = "https://github.com/psf/black-pre-commit", rev = "24.4.0" },
    { repo = "https://github.com/pycqa/isort", rev = "5.13.2" },
]